reorganized code and create installation package
