+++ regparse/plugins/activesetup.plugin
