+++ regparse/plugins/appinit.plugin
