+++ regparse/plugins/bhos.plugin
