+++ regparse/plugins/knowndlls.plugin
