+++ regparse/plugins/searchit.plugin
