+++ regparse/plugins/sysinfo.plugin
