+++ regparse/plugins/sysinfo.py
@@ -1,8 +1,11 @@
 import sys
 import time
+
 from Registry import Registry
-from PluginManager import HelperFunctions
-from jinja2 import Template, Environment, PackageLoader
+from jinja2 import Environment, PackageLoader
+
+from regparse.PluginManager import HelperFunctions
+
 
 class PluginClass(object):
