+++ regparse/plugins/typedpaths.plugin
