+++ regparse/plugins/usbstor.plugin
