+++ regparse/plugins/winlogon.plugin
