+++ regparse/plugins/winlogon.py
